Version Entrega 3 8/07/2023

+++ alumnos/views.py
@@ -1,9 +1,15 @@
 from django.shortcuts import render,redirect 
 from .models import Cliente 
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import logout
+from django.http import HttpResponse
 
 # Create your views here.
 
+def home(request):
+    context ={}
+    return render(request, 'alumnos/home.html')
 
 def Inicio(request):
     context ={}
@@ -25,29 +31,38 @@ def suscribirse(request):
     context ={}
     return render(request, 'alumnos/suscribirse.html')
 
+@login_required
 def gestion(request):
     listadoClientes = Cliente.objects.all()
     return render(request, 'alumnos/gestion.html', {"clientes" : listadoClientes })
 
+def exit(request):
+    logout(request)
+    return redirect('login')
+
+@login_required
 def registrar(request):
     rut=request.POST['txtRut']
     nombre=request.POST['txtNombre']
     apellidoP=request.POST['txtApellidoP']
     apellidoM=request.POST['txtApellidoM']
 
+
     cliente = Cliente.objects.create(
         rut=rut , nombre=nombre, apellido_paterno=apellidoP, apellido_materno=apellidoM)
+    if not request.user.is_superuser:
+        return HttpResponse("No tienes permisos para editar este usuario.")
+    return render(request,'alumnos/edicion.html', {"cliente": cliente})
     return redirect('gestion')
 
+@login_required
 def edicion(request, rut):
     cliente = Cliente.objects.get(rut = rut)
-    data={
-
-
-
-    }
+    if not request.user.is_superuser:
+        return HttpResponse("No tienes permisos para editar este usuario.")
     return render(request,'alumnos/edicion.html', {"cliente": cliente})
 
+@login_required
 def editar(request):
     rut = request.POST['rut']
     nombre = request.POST['nombre']
@@ -62,10 +77,14 @@ def editar(request):
     cliente.save()
     return redirect('gestion')
 
-
-
+@login_required
 def eliminar(request, rut):
     cliente= Cliente.objects.get(rut=rut)
+    if not request.user.is_superuser:
+        return HttpResponse("No tienes permisos para editar este usuario.")
+    return render(request,'alumnos/edicion.html', {"cliente": cliente})
+
+    
     cliente.delete()
     return redirect('gestion')
